requirements-dev.txt
pytest==8.3.5
pytest-django==4.9.0
factory_boy==3.3.1
black==24.10.0
isort==5.13.2
flake8==7.1.1
